Extract ancestry logic to standalone module
